+++ src/com/aixcoder/core/Learner.java
@@ -3,7 +3,6 @@ package com.aixcoder.core;
 import java.io.File;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
-import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;