+++ CHANGELOG.md
@@ -1,12 +1,43 @@
 # Change Log
 All notable changes to the "aiXcoder" extension will be documented in this file.
 
+## [1.2.0] - 2019-12-31
+
+### Change
+- Merged online and local branch.
+
+## [1.1.9] - 2019-12-16
+### Fix
+- Fix a bug that make service start multiple times.
+
+## [1.1.8] - 2019-12-11
+### Added
+- Show local service indexing status.
+
+### Fix
+- Fix a bug that make service failed to start.
+
+
+## [1.1.6] - 2019-11-21
+### Fix
+- (Mac) Execute permission issue
+
 ## [1.1.5] - 2019-11-21
 
 ### Add
 ### Change
 ### Fix
 - Java 1.6 compliance
+
+## [1.1.4] - 2019-11-19
+### Add
+- Now works better with project-scope completion.
+
+### Change
+
+### Fix
+- Fix a bug when there is quote in comment.
+- Kill local service when updating to prevent file overwrite failures.
 
 ## [1.1.3] - 2019-11-04
 ### 新增